+++ src/components/eholdings/eholdings.test.js
@@ -12,7 +12,7 @@ jest.mock('../../routes/provider-edit', () => () => <span>ProviderEdit</span>);
 jest.mock('../../routes/package-show', () => () => <span>PackageShow</span>);
 jest.mock('../../routes/package-edit', () => () => <span>PackageEdit</span>);
 jest.mock('../../routes/package-create', () => () => <span>PackageCreate</span>);
-jest.mock('../../routes/title-show', () => () => <span>TitleShow</span>);
+jest.mock('../../routes/title-show-route', () => () => <span>TitleShowRoute</span>);
 jest.mock('../../routes/title-edit-route', () => () => <span>TitleEditRoute</span>);
 jest.mock('../../routes/title-create-route', () => () => <span>TitleCreateRoute</span>);
 jest.mock('../../routes/resource-show', () => () => <span>ResourceShow</span>);